requirements-airflow.txt
# Apache Airflow
apache-airflow==2.8.0
apache-airflow-providers-apache-spark==4.7.1
apache-airflow-providers-celery==3.5.1

# MLflow & ML Libraries
mlflow==2.9.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Additional Dependencies
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.34.0
pyarrow==14.0.0
